requirements.txt
fastapi==0.81.0
SQLAlchemy==1.4.40
uvicorn==0.18.3
jinja2==3.1.2
psycopg2-binary==2.9.5
python-multipart
python-jose[cryptography]
passlib[bcrypt]